refactor: 운동 생성 API에서 중첩 record 패턴 적용
