+++ src/main/java/umc/cockple/demo/domain/exercise/domain/Exercise.java
@@ -3,9 +3,8 @@ package umc.cockple.demo.domain.exercise.domain;
 import jakarta.persistence.*;
 import lombok.*;
 import umc.cockple.demo.domain.exercise.dto.ExerciseAddrUpdateCommand;
-import umc.cockple.demo.domain.exercise.dto.ExerciseCreateCommand;
+import umc.cockple.demo.domain.exercise.dto.ExerciseCreateDTO;
 import umc.cockple.demo.domain.exercise.dto.ExerciseUpdateCommand;
-import umc.cockple.demo.domain.exercise.dto.GuestInviteCommand;
 import umc.cockple.demo.domain.member.domain.MemberExercise;
 import umc.cockple.demo.domain.party.domain.Party;
 import umc.cockple.demo.global.common.BaseEntity;
@@ -65,7 +64,7 @@ public class Exercise extends BaseEntity {
     @Builder.Default
     private List<Guest> guests = new ArrayList<>();
 
-    public static Exercise create(ExerciseAddr exerciseAddr, ExerciseCreateCommand command) {
+    public static Exercise create(ExerciseAddr exerciseAddr, ExerciseCreateDTO.Command command) {
         return Exercise.builder()
                 .exerciseAddr(exerciseAddr)
                 .date(command.date())
