--- src/main/java/umc/cockple/demo/domain/exercise/dto/ExerciseAddrCreateCommand.java
@@ -1,12 +1,0 @@
-package umc.cockple.demo.domain.exercise.dto;
-
-import lombok.Builder;
-
-@Builder
-public record ExerciseAddrCreateCommand(
-        String roadAddress,
-        String buildingName,
-        Double latitude,
-        Double longitude
-) {
-}
