--- src/main/java/umc/cockple/demo/domain/exercise/dto/ExerciseCreateCommand.java
@@ -1,18 +1,0 @@
-package umc.cockple.demo.domain.exercise.dto;
-
-import lombok.Builder;
-
-import java.time.LocalDate;
-import java.time.LocalTime;
-
-@Builder
-public record ExerciseCreateCommand(
-        LocalDate date,
-        LocalTime startTime,
-        LocalTime endTime,
-        Integer maxCapacity,
-        Boolean partyGuestAccept,
-        Boolean outsideGuestAccept,
-        String notice
-) {
-}
