--- src/main/java/umc/cockple/demo/domain/exercise/dto/ExerciseCreateResponseDTO.java
@@ -1,12 +1,0 @@
-package umc.cockple.demo.domain.exercise.dto;
-
-import lombok.Builder;
-
-import java.time.LocalDateTime;
-
-@Builder
-public record ExerciseCreateResponseDTO(
-        Long exerciseId,
-        LocalDateTime createdAt
-) {
-}
