+++ src/main/java/umc/cockple/demo/domain/exercise/dto/ExerciseJoinDTO.java
@@ -1,0 +1,4 @@
+package umc.cockple.demo.domain.exercise.dto;
+
+public class ExerciseJoinDTO {
+}
