--- src/main/java/umc/cockple/demo/domain/exercise/dto/ExerciseJoinResponseDTO.java
@@ -1,14 +1,0 @@
-package umc.cockple.demo.domain.exercise.dto;
-
-import lombok.Builder;
-
-import java.time.LocalDateTime;
-
-@Builder
-public record ExerciseJoinResponseDTO(
-        Long participantId,
-        Integer participantNumber,
-        LocalDateTime joinedAt,
-        Integer currentParticipants
-) {
-}
